refactor common tests using pytest parametrize
