+++ tests/tests_imputer/conftest.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from collections.abc import Callable
+
 from typing import Any
 
 import numpy as np
