+++ tests/tests_imputer/test_gaussian.py
@@ -8,84 +8,7 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from shapiq_student.imputer.gaussian.exceptions import CategoricalFeatureError
 from shapiq_student.imputer.gaussian.gaussian_imputer import GaussianImputer
-
-
-class TestInputValidation:
-    """Tests that input data is handled correctly and that malformed data is detected successfully."""
-
-    def test_categorical_feature_check_only_continuous(self, dummy_model) -> None:
-        """Test that no error is raised when all features are continuous with >2 unique values."""
-        data = np.array(
-            [
-                [1.0, 2.0, 3.0],
-                [4.0, 5.0, 6.0],
-                [7.0, 8.0, 9.0],
-            ]
-        )
-        x = np.array([[2.0, 3.0, 4.0]])
-        # Should not raise error
-        GaussianImputer(model=dummy_model, data=data, x=x)
-
-    def test_categorical_feature_check_binary_column(self, dummy_model) -> None:
-        """Test that an exception is raised if the background data has column containing only two unique values."""
-        data = np.array(
-            [
-                [1.0, 0, 3.0],
-                [2.0, 1, 4.0],
-                [3.0, 0, 5.0],
-            ]
-        )
-        x = np.array([2.0, 1.0, 4.0])
-        with pytest.raises(CategoricalFeatureError) as exc:
-            GaussianImputer(model=dummy_model, data=data, x=x)
-        msg = str(exc.value)
-        # The second column (index 1) should be flagged as categorical, so 'f2' should be in the message
-        assert "f2" in msg
-        # The first and third columns should not be flagged, so 'f1' and 'f3' should not be in the message
-        assert "f1" not in msg
-        assert "f3" not in msg
-
-    def test_categorical_feature_check_string(self, dummy_model) -> None:
-        """Test that an exception is raised if the background data has a column containing string values."""
-        data = np.array(
-            [
-                [1.0, "a", 3.0],
-                [2.0, "b", 4.0],
-                [3.0, "a", 5.0],
-            ],
-            dtype=object,
-        )
-        x = np.array([2.0, "b", 4.0], dtype=object)
-        with pytest.raises(CategoricalFeatureError) as exc:
-            GaussianImputer(model=dummy_model, data=data, x=x)
-        msg = str(exc.value)
-        assert "f2" in msg
-        assert "f1" not in msg
-        assert "f3" not in msg
-
-    def test_categorical_feature_check_mixed(self, dummy_model) -> None:
-        """Test that an exception is raised if the background data has binary and string-valued columns."""
-        data = np.array(
-            [
-                [1.0, 0, "a", 3.0],
-                [2.0, 1, "b", 4.0],
-                [3.0, 0, "a", 5.0],
-            ],
-            dtype=object,
-        )
-        x = np.array([2.0, 1.0, "b", 4.0], dtype=object)
-        with pytest.raises(CategoricalFeatureError) as exc:
-            GaussianImputer(model=dummy_model, data=data, x=x)
-        msg = str(exc.value)
-        # Both f2 (index 1) and f3 (index 2) must be mentioned
-        assert "f2" in msg
-        assert "f3" in msg
-        # f1 and f4 must not appear
-        assert "f1" not in msg
-        assert "f4" not in msg
-
 
 ##############################################
 # Tests for Cov Mat and Mean Calculation --- #
